requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
motor==3.4.0
pymongo==4.7.2
python-multipart==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
APScheduler==3.10.4
slowapi==0.1.9
requests==2.31.0
twilio==9.1.0
sendgrid==6.11.0
python-dotenv==1.0.1
email-validator==2.3.0
pytz==2026.1
httpx==0.27.2
python-json-logger==2.0.7
cryptography==46.0.6
